+++ src/util/common.c
@@ -1,3 +1,3 @@
 #include "common.h"
 
-extern inline i32 max_i32(i32 a, i32 b);
+extern inline i64 max_i32(i64 a, i64 b);
